REAL Initial Commit
Our first real commit to the repo! This is the mostly functional app.
The only thing that doesn't currently work is the Help button, which I'm
not sure exactly what I plan to do with. However, the app runs in both
"setup" (GUI) mode and "batch" (cron/scheduled task) mode. Run the GUI
mode to set your settings, including your API key and to choose your
domains. The click Update Now to do DNS update interactively, or
schedule it as a automated job by specifying "-run" on the command line.
Exception handling is pretty robust, but many error boxes just show the
raw exceptions. I plan to edit the README soon to give a better overview
of what this does and how to use it.