Fix for Linode API changes
It looks like Linode has reordered some elements of the response JSON,
swapping the location of the ERRORARRAY and ACTION keywords. (ERRORARRAY
was at the beginning and ACTION at the front; these are now reversed.)
Admittedly, hard-coding the locations of these probably isn't a good
idea; it would probably be better to parse the JSON and look for the
existence of an empty ERRORARRAY, etc. That said, I didn't want to drag
in a full-on JSON parser into this mix and tried to shortcut things by
doing some simple string tests. No idea how long this has been "broken".
Linode's API docs on their site still have ERRORARRAY appearing first. I
only noticed this when I found my local copy of the program wasn't
updating anymore.